% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dat_TA}
\alias{dat_TA}
\title{Sapwood statistics from Sweden}
\format{
A data frame with 148 rows and 4 variables:
\describe{
  \item{H}{Number of heartwood rings}
  \item{S}{Number of sapwood rings}
  \item{W}{Mean tree ring width of heartwood}
}
}
\source{
<https://github.com/karirogg/ScotsPine>
}
\usage{
dat_TA
}
\description{
A dataset of sapwood statistics from historical samples in southern Sweden
}
\keyword{datasets}
