% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sapwood_utils.R
\name{plot.sapwood_fit}
\alias{plot.sapwood_fit}
\title{Plot method for sapwood models}
\usage{
\method{plot}{sapwood_fit}(
  object,
  type = "fit",
  xlim = NULL,
  ylim = NULL,
  prediction = T,
  confidence = F
)
}
\arguments{
\item{object}{an object of class "sapwood_fit", a result from a call to \code{sapwood_fit_l}, \code{sapwood_fit_pl} or \code{sapwood_fit_plw}.}

\item{type}{Type of plot. Possible types are:
\itemize{
\item{"fit"}{Plots sapwood rings versus heartwood rings and the best fit, including prediction and/or confidence intervals for the median (if wanted). Not available for an object of type "parabolic_linear_W' (result from \code{sapwood_fit_plw})}
\item{"residual"}{Plots standardized residuals on log scale of the model as well as dotted lines for z_0.025 and z_0.975}
\item{"qq"}{Plots a QQ plot of the standardized residuals}
}}

\item{xlim, ylim}{Limits of the axes of the plot. Given as a vector similar to c(0,100). \code{xlim} defaults to the range from 0 to the maximum value of heartwood in the data (with some padding). \code{ylim} defaults to 0 to the maximum value of predicted data with some padding.}

\item{prediction}{If TRUE, prediction bands are plotted on the plot. Only used for type = "fit".}

\item{confidence}{If TRUE, confidence bands to the median are plotted to the plot. Only used for type = "fit".}
}
\description{
Visualize sapwood model.
}
\examples{
data(dat_TA)
fit <- sapwood_fit_pl(S~H, dat_TA)
plot(fit, xlim=c(0,220), ylim=c(0,150))
plot(fit)
plot(fit, type="residual")
plot(fit, type="qq")
}
