% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sapwood_fit.R
\name{sapwood_fit}
\alias{sapwood_fit}
\alias{sapwood_fit_l}
\alias{sapwood_fit_pl}
\alias{sapwood_fit_plw}
\title{A model for sapwood rings in scots pine}
\usage{
sapwood_fit_l(
  formula,
  dat,
  alpha = 0.05,
  mu_theta_1 = log(0.1),
  sd_theta_1 = (log(0.1) + 4 * log(10))/2,
  mu_theta_2 = log(0.1),
  sd_theta_2 = (log(0.1) + 4 * log(10))/2,
  H_0 = 100
)

sapwood_fit_pl(
  formula,
  dat,
  alpha = 0.05,
  mu_theta_1 = log(0.1),
  sd_theta_1 = (log(0.1) + 4 * log(10))/2,
  mu_theta_2 = log(0.1),
  sd_theta_2 = (log(0.1) + 4 * log(10))/2,
  H_0 = 100
)

sapwood_fit_plw(
  formula,
  dat,
  alpha = 0.05,
  mu_theta_1 = log(0.1),
  sd_theta_1 = (log(0.1) + 4 * log(10))/2,
  mu_theta_2 = log(0.1),
  sd_theta_2 = (log(0.1) + 4 * log(10))/2,
  H_0 = 100
)
}
\arguments{
\item{formula}{Formula for the fit. If using \code{sapwood_fit_pl} or \code{sapwood_fit_l},
the formula should be on the form \code{S~H}, where S is the name of the column containing the
number of sapwood rings and H is the name of the column containing the number of heartwood rings.
If using \code{sapwood_fit_plw}, the formula should be on the form \code{S~H+W}, where S and H are as before,
and W is the name of the column including data on mean tree ring width.}

\item{dat}{Dataset to be fitted to. Column names should match formula arguments}

\item{alpha}{Confidence of the fit, for prediction, confidence for median and parameter confidence intervals. defaults to 0.05 (which corresponds to 95\% confidence)}

\item{mu_theta_1}{Regularization mu for theta_1 (see Edvardsson et al. 2021)}

\item{sd_theta_1}{Regularization sigma for theta_1 (see Edvardsson et al. 2021)}

\item{mu_theta_2}{Regularization mu for theta_2 (see Edvardsson et al. 2021)}

\item{sd_theta_2}{Regularization sigma for theta_2 (see Edvardsson et al. 2021)}

\item{H_0}{Cutoff point for parabolic-linear model. Only relevant for \code{sapwood_fit_pl} and \code{sapwood_fit_plw}.}
}
\value{
The functions return objects of class "sapwood_fit". An object of class "plm" is a list containing the following components:\cr
\item{\code{parameter_CI}}{Confidence intervals for the parameters of the model.}
\item{\code{predictions}}{Predictions for the model for H between 0 and 500, if \code{sapwood_fit_l} or \code{sapwood_fit_pl} are used, prediction and confidence intervals for the fit is included. If \code{sapwood_fit_plw} is used, only median of the prediction is returned.}
\item{\code{residuals}}{Standardized residuals of the model. See Edvardsson et al. 2021.}
\item{\code{AIC}}{AIC of the model}
\item{\code{formula}}{The formula for the model, as specified in the input.}
\item{\code{type}}{Type of the model, "parabolic_linear_W", "parabolic_linear" or "linear" (Models 1,2 and 3, respectively).}
\item{\code{alpha}}{Confidence of the model}
}
\description{
The function takes in heartwood and sapwood (and possibly tree ring width) data from a dataset and returns a fit and some information about the fit, such as prediction intervals, figures, confidence intervals for parameters etc. See Edvardsson et al.
}
\examples{
data(dat_TA)
fit <- sapwood_fit_pl(S~H, dat_TA)
plot(fit)
plot(fit, type="residual")
plot(fit, type="qq")
residuals(fit)
AIC(fit)

}
