% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smaland}
\alias{smaland}
\title{Sapwood statistics from Sweden}
\format{
A data frame with 268 rows and 3 variables:
\describe{
  \item{H}{Number of heartwood rings}
  \item{W}{Mean tree ring width of heartwood}
  \item{S}{Number of sapwood rings}
}
}
\source{
<https://github.com/karirogg/SapwoodModel>
}
\usage{
smaland
}
\description{
A dataset of sapwood statistics from historical samples in Småland, southern Sweden
}
\keyword{datasets}
