% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sapwood_utils.R
\name{summary.sapwood_fit}
\alias{summary.sapwood_fit}
\title{Summary method for sapwood fits}
\usage{
\method{summary}{sapwood_fit}(object)
}
\arguments{
\item{object}{an object of class "sapwood_fit", a result from a call to \code{sapwood_fit_l}, \code{sapwood_fit_pl} or \code{sapwood_fit_plw}.}
}
\description{
\code{summary} method for class "sapwood_fit"
}
