% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uppland}
\alias{uppland}
\title{Sapwood statistics from Sweden}
\format{
A data frame with 47 rows and 3 variables:
\describe{
  \item{H}{Number of heartwood rings}
  \item{S}{Number of sapwood rings}
  \item{W}{Mean tree ring width of heartwood}
}
}
\source{
<https://github.com/karirogg/SapwoodModel>
}
\usage{
uppland
}
\description{
A dataset of sapwood statistics from living samples from Uppland, Sweden
}
\keyword{datasets}
