% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sapwood_utils.R
\name{predict.sapwood_fit}
\alias{predict.sapwood_fit}
\title{Prediction for sapwood rings}
\usage{
\method{predict}{sapwood_fit}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{an object of class "sapwood_fit", a result from a call to \code{sapwood_fit_l}, \code{sapwood_fit_pl} or \code{sapwood_fit_plw}.}

\item{newdata}{an optional data frame/tibble/vector in which to look for variables with which to predict. If omitted, the fitted values are used.}
}
\description{
Obtain predictions for a model of type 'sapwood_fit',
including prediction and confidence intervals (for \code{sapwood_fit_l} and \code{sapwood_fit_pl}).
For an object resulted in a \code{sapwood_fit_plw} call, only prediction median is included.
}
