% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sapwood_utils.R
\name{predict.sapwood_fit}
\alias{predict.sapwood_fit}
\title{Prediction for sapwood rings}
\usage{
\method{predict}{sapwood_fit}(object, newdata = NULL, confidence = 0.95, ...)
}
\arguments{
\item{object}{an object of class "sapwood_fit", a result from a call to \code{sapwood_fit_l}, \code{sapwood_fit_pl} or \code{sapwood_fit_plw}.}

\item{newdata}{an optional data frame/tibble/vector in which to look for variables with which to predict. If omitted, the fitted values are used. If a column \code{remaining} is in the \code{newdata} tibble/data frame, it will be treated as remaining sapwood rings and the prediction interval will be based on that}
}
\description{
Obtain predictions for a model of type 'sapwood_fit',
including prediction and confidence intervals.
Optionally, one can predict how many sapwood rings were in a sample
that contains some remaining rings but not all.
}
